+++ RELEASE_NOTES.md
@@ -12,6 +12,19 @@ While the new TargetCategory class supports subtypes, only reading them is curre
   * `TargetIds(ComponentIds(1), ComponentIds(2), ComponentIds(3))`
  * `TargetCategories` can be used to specify one or more target categories:
   * `TargetCategories(ComponentCategory.BATTERY, ComponentCategory.INVERTER)`
+* `DispatchApiClient.stream()` now returns a streamer instead of a receiver. This allows you to use the new `new_receiver` method to create a receiver with more options, such as `max_size`, `warn_on_overflow` and `include_events`.
+
+    ```python
+    client = DispatchApiClient(
+        key="key",
+        server_url="grpc://dispatch.url.goes.here.example.com"
+    )
+    receiver = client.stream(microgrid_id=1).new_receiver()
+
+    async for message in receiver:
+        print(message.event, message.dispatch)
+    ```
+
 
 ## New Features